facebook/requirements.txt
# Facebook API
requests>=2.31.0

# Google APIs
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Cloud Functions Framework
functions-framework>=3.4.0

# Date handling
python-dateutil>=2.8.2